+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 `Tabman` adheres to [Semantic Versioning](http://semver.org/).
 
 #### 2.x Releases
-- `2.2.x` Releases - [2.2.0](#220)
+- `2.2.x` Releases - [2.2.0](#220) | [2.2.1](#221)
 - `2.1.x` Releases - [2.1.0](#210) | [2.1.1](#211) | [2.1.2](#212) | [2.1.3](#213) | [2.1.4](#214)
 - `2.0.x` Releases - [2.0.0](#200)
 
@@ -28,9 +28,14 @@ All notable changes to this project will be documented in this file.
 - `0.4.x` Releases - [0.4.0](#040) | [0.4.1](#041) | [0.4.2](#042) | [0.4.3](#043) | [0.4.4](#044) | [0.4.5](#045) | [0.4.6](#046) | [0.4.7](#047) | [0.4.8](#048)
 
 ---
+## [2.2.0](https://github.com/uias/Tabman/releases/tag/2.2.1)
+Released on 2019-03-24
+
+#### Updated
+- AutoInsetter to 1.6 - with numerous fixes and improvements for view controller insetting.
 
 ## [2.2.0](https://github.com/uias/Tabman/releases/tag/2.2.0)
-Released on 2019-01-14
+Released on 2019-03-14
 
 #### Added 
 - [#363](https://github.com/uias/Tabman/issues/363) `.alignment` property to `TMBarLayout` to adjust content alignment.